+++ src/Components/CardSkeleton.js
@@ -1,9 +1,24 @@
 import styled from 'styled-components';
+import { Shimmer } from 'Common';
+
+const CardSkeleton = () => {
+  return (
+    <CardBox>
+      <ImageWrap>
+        <Shimmer height="200px" margin="0" />
+      </ImageWrap>
+      <ContentWrap>
+        <Shimmer width="100px" />
+        <Shimmer />
+      </ContentWrap>
+    </CardBox>
+  );
+};
 
 const CardBox = styled.div`
   width: 19rem;
-  height: 364px;
-  margin: 1rem;
+  height: 333px;
+  margin: -1rem 1rem;
   position: relative;
   display: flex;
   flex-wrap: wrap;
@@ -17,31 +32,25 @@ const CardBox = styled.div`
     width: 100%;
     margin: 0px 0px 1rem 0px;
   }
-
   @media (max-width: 944px) {
     width: calc(50% - 2rem);
   }
-
   @media (min-width: 944px) {
     &:hover {
       transform: translateY(-10px);
       box-shadow: rgba(0, 0, 0, 0.1) 0px 4px 16px 0px;
     }
   }
-
   @media (max-width: 944px) {
     width: 487px;
     height: 464px;
   }
-
   @media (max-width: 1056px) {
     width: 42vw;
   }
-
   @media (max-width: 767px) {
     width: 92vw;
   }
-
   @media (max-width: 500px) {
     width: 451px;
   }
@@ -60,47 +69,4 @@ const ContentWrap = styled.div`
   height: 165px;
 `;
 
-const ContentLineFirst = styled.div`
-  width: 100px;
-  height: 20px;
-  margin: 20px 20px 0 20px;
-  background-color: #eeeeee;
-  overflow: hidden;
-`;
-
-const ContentLine = styled.div`
-  width: 200px;
-  height: 20px;
-  margin: 20px 20px 0 20px;
-  background-color: #eeeeee;
-  overflow: hidden;
-`;
-
-const Shimmer = styled.div`
-  width: 40%;
-  height: 100%;
-  background-color: #e0e0e0;
-  box-shadow: 0 0 30px 30px #e0e0e0;
-  animation: loading 1s infinite;
-
-  @keyframes loading {
-    0% {
-      transform: translateX(-50%);
-    }
-    50% {
-      transform: translateX(100%);
-    }
-    100% {
-      transform: translate(200%);
-    }
-  }
-`;
-
-export const style = {
-  CardBox,
-  ImageWrap,
-  ContentLine,
-  ContentLineFirst,
-  ContentWrap,
-  Shimmer,
-};
+export default CardSkeleton;
