--- src/Pages/ListPage/ListPageStyle.js
@@ -1,45 +1,0 @@
-import styled from 'styled-components';
-
-const Wrapper = styled.div`
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  background: #f8f9fa;
-`;
-
-const Container = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  margin-top: 2rem;
-  width: 1728px;
-  margin-left: auto;
-  margin-right: auto;
-  background: #f8f9fa;
-
-  @media (max-width: 1919px) {
-    width: 1376px;
-  }
-  @media (max-width: 1440px) {
-    width: 1124px;
-  }
-  @media (max-width: 1056px) {
-    width: calc(80% - 2rem);
-    margin-left: 1rem;
-    margin-right: 1rem;
-  }
-  @media (max-width: 767px) {
-    width: calc(100% - 2rem);
-  }
-`;
-
-const CardList = styled.div`
-  width: 100%;
-  display: flex;
-  flex-wrap: wrap;
-  align-items: stretch;
-  justify-content: left;
-  background: #f8f9fa;
-`;
-
-export const style = { Wrapper, Container, CardList };
