+++ src/Routes.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
-import ListPage from 'Pages/ListPage/ListPage';
+import ListPage from 'Pages/ListPage';
 import WritePage from 'Pages/WritePage/WritePage';
 import DetailPage from 'Pages/DetailPage/DetailPage';
-import EditPage from 'Pages/EditPage/EditPage';
+import EditPage from 'Pages/EditPage';
 
 const Routes = () => {
   return (
